+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Support for cascaded WMS layers [#77](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/77)
 * Support for 'GeoGig' GeoServer store/repository types [#76](https://gitlab.data.bas.ac.uk/MAGIC/web-map-inventory/issues/76)
-
 
 ### Fixed